backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
sentence-transformers
faiss-cpu
llama-cpp-python==0.2.75
pypdf==3.17.4
python-docx==1.1.0
numpy==1.26.3

huggingface-hub

networkx
